Move body generator for tuple structs to its own module
Cut-paste
